+++ test/integration/associations/has-one.test.js
@@ -530,8 +530,8 @@ describe(Support.getTestDialectTeaser('HasOne'), function() {
         });
 
         expect(Profile.rawAttributes.uid).to.be.ok;
-        expect(Profile.rawAttributes.uid.references).to.equal(User.getTableName());
-        expect(Profile.rawAttributes.uid.referencesKey).to.equal('id');
+        expect(Profile.rawAttributes.uid.references.model).to.equal(User.getTableName());
+        expect(Profile.rawAttributes.uid.references.key).to.equal('id');
         expect(Profile.rawAttributes.uid.allowNull).to.be.false;
 
         // Let's clear it
@@ -544,8 +544,8 @@ describe(Support.getTestDialectTeaser('HasOne'), function() {
         });
 
         expect(Profile.rawAttributes.uid).to.be.ok;
-        expect(Profile.rawAttributes.uid.references).to.equal(User.getTableName());
-        expect(Profile.rawAttributes.uid.referencesKey).to.equal('id');
+        expect(Profile.rawAttributes.uid.references.model).to.equal(User.getTableName());
+        expect(Profile.rawAttributes.uid.references.key).to.equal('id');
         expect(Profile.rawAttributes.uid.allowNull).to.be.false;
       });
 
@@ -566,8 +566,8 @@ describe(Support.getTestDialectTeaser('HasOne'), function() {
         User.hasOne(Profile, { foreignKey: Profile.rawAttributes.user_id});
 
         expect(Profile.rawAttributes.user_id).to.be.ok;
-        expect(Profile.rawAttributes.user_id.references).to.equal(User.getTableName());
-        expect(Profile.rawAttributes.user_id.referencesKey).to.equal('uid');
+        expect(Profile.rawAttributes.user_id.references.model).to.equal(User.getTableName());
+        expect(Profile.rawAttributes.user_id.references.key).to.equal('uid');
         expect(Profile.rawAttributes.user_id.allowNull).to.be.false;
       });
 
@@ -589,8 +589,8 @@ describe(Support.getTestDialectTeaser('HasOne'), function() {
 
         expect(Project.rawAttributes.userUid).to.be.ok;
         expect(Project.rawAttributes.userUid.allowNull).to.be.false;
-        expect(Project.rawAttributes.userUid.references).to.equal(User.getTableName());
-        expect(Project.rawAttributes.userUid.referencesKey).to.equal('uid');
+        expect(Project.rawAttributes.userUid.references.model).to.equal(User.getTableName());
+        expect(Project.rawAttributes.userUid.references.key).to.equal('uid');
         expect(Project.rawAttributes.userUid.defaultValue).to.equal(42);
       });
     });