+++ test/integration/sequelize.test.js
@@ -737,7 +737,7 @@ describe(Support.getTestDialectTeaser('Sequelize'), function() {
 
       it('returns an error correctly if unable to sync a foreign key referenced model', function() {
         this.sequelize.define('Application', {
-          authorID: { type: Sequelize.BIGINT, allowNull: false, references: 'User', referencesKey: 'id' }
+          authorID: { type: Sequelize.BIGINT, allowNull: false, references: { model: 'User', key: 'id' } }
         });
 
         return this.sequelize.sync().catch(function(error) {